Ability to evaluate a python function
Can now pass a python function to TestProgram instead of a script.  The
function is evaluated using a process pool from the multiprocessing
package. This greatly speeds things up. The old way of passing a script
file is still available too.

Many changes to hosts.py to support this function.